archive/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Streamlit Frontend
streamlit==1.28.1
requests==2.31.0

# LangChain & LangGraph (will be added in Sprint 2)
# langchain==0.1.17
# langgraph==0.0.55
# langchain-community==0.0.38

# AI/LLM (will be added in Sprint 2)
# anthropic==0.25.0

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pandas==2.1.4

# File Export
fpdf2==2.7.6

# Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0
uuid==1.30